mpir: split lpid headers to mpir_lpid.h
Because we need access MPIR_Lpid definitions in mpidpre headers, we need
move worlds and lpid definitions to device-independent headers.

Add macro MPIR_LPID_INVALID.

Make MPIR_Lpid signed. Since we are going to perform arithmetic on
MPIR_Lpid, e.g. in using strided pmap, make MPIR_Lpid int64_t instead of
uint64_t to avoid accidental conversion errors.
